refactor(assemblies): extract config to separate file for better maintainability
Move the configuration object (mod and fields) from Assemblies.tsx to a dedicated config file (assemblies.config.tsx). This improves code organization, separates concerns, and makes the main component cleaner and easier to read. Use refs to handle circular dependency between the config function and the useCrud hook.
